+++ synapseclient/extensions/curator/get_grid.py
